requirements.txt
fastapi
uvicorn
gradio
langchain
sentence-transformers
faiss-cpu
transformers
torch
pandas
sqlalchemy
pillow
pytest
python-dotenv